correct the email address validation regex
A number of characters (including, particularly "-") are legal in email addresses but not included in the original regex. Also, the regex was not anchored, so in many cases where an email included a character not allowed in the regex, it would still pass as a partial match, thus possibly faking out tests in a confusing way. And of course that would also allow many illegal addresses to pass.