Fixed render_header_value eternal loop if a value of nil is present
In general this bug doesn't happen if rendering an entire message, since
render_headers detects and drops nils.

However if one should render a header directly using render_header it
bypasses this reduction and ends up looping in render_header_value

This happens since it falls back to the last overload of the function,
which wraps the value into a list.

Since render_header_value can't handle empty lists, it falls back again
and wraps the empty list (producing the same list), and keeps looping.

The change presented here will treat such cases as a critical failure
and will crash.

This is chosen instead of quietly ignoring the issue or casting the nil
to a string "nil" to avoid hiding potential errors